drivers: misc: mathworks: fix for copy_to_user() calls
The kernel has some new compile-time checks to make sure that the size of
source and destination params are identical.

These fail on ARM64, since `unsigned int` and `unsigned long` differ in
length.

+++ drivers/misc/mathworks/mw_stream_channel.c
@@ -338,8 +338,8 @@ static long mwadma_rx_ctl(struct mwadma_dev *mwdev, unsigned int cmd, unsigned l
     struct mw_axidma_params usrbuf;
     struct mwadma_chan *mwchan = mwdev->rx;
     enum mwadma_chan_status     status;
-    unsigned int                next_index, done_index;
-    unsigned int                error;
+    unsigned long               next_index, done_index;
+    unsigned long               error;
     unsigned long int flags;
     struct mwadma_slist *tmp = NULL;
     switch(cmd)
